Assessment Rest framework API/requirement.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
future==1.0.0
pytz==2024.1
serializers==0.2.4
sqlparse==0.4.4
tzdata==2024.1
